+++ src/dapps/MakerSai/components/SelectCdpEntry/SelectCdpEntry.vue
@@ -124,7 +124,7 @@ export default {
     return {};
   },
   computed: {
-    ...mapState(['account', 'gasPrice', 'web3', 'network', 'ens']),
+    ...mapState('main', ['account', 'gasPrice', 'web3', 'network', 'ens']),
     hasProxy() {
       if (this.aCdp) {
         return !this.aCdp.noProxy;