+++ src/dapps/MakerSai/containers/SelectCDP/SelectCDP.vue
@@ -94,7 +94,7 @@ export default {
     };
   },
   computed: {
-    ...mapState(['account', 'gasPrice', 'web3', 'network', 'ens']),
+    ...mapState('main', ['account', 'gasPrice', 'web3', 'network', 'ens']),
     cdpOptions() {
       return this.availableCdps;
     }