+++ src/helpers/misc.js
@@ -166,7 +166,7 @@ const solidityType = inputType => {
 };
 
 const isDarklisted = addr => {
-  const storedDarklist = store.state.darklist.data;
+  const storedDarklist = store.state.main.darklist.data;
   const darklisted =
     storedDarklist > 0
       ? storedDarklist.findIndex(item => {
@@ -177,7 +177,7 @@ const isDarklisted = addr => {
         })
       : -1;
   const errMsg =
-    darklisted === -1 ? '' : store.state.darklist.data[darklisted].comment;
+    darklisted === -1 ? '' : store.state.main.darklist.data[darklisted].comment;
   const errObject = {
     error: darklisted === -1 ? false : true,
     msg: errMsg