+++ src/layouts/InterfaceLayout/containers/InteractWithContractContainer/InteractWithContractContainer.vue
@@ -295,7 +295,7 @@ export default {
     };
   },
   computed: {
-    ...mapState(['network', 'gasPrice', 'account', 'web3']),
+    ...mapState('main', ['network', 'gasPrice', 'account', 'web3']),
     mergedContracts() {
       const customContracts = store.get('customContracts') || [];
       const concatContracts = this.network.type.contracts.concat(