+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 ### Feature
+
 - ENS functionality for address book [#1888] (https://github.com/MyEtherWallet/MyEtherWallet/pull/1888)
+
+### Bug
+
+- Set ENS in offline helper [#1891](https://github.com/MyEtherWallet/MyEtherWallet/pull/1891)
 
 ### Devop
 
+- Modularize store state [#1722][https://github.com/MyEtherWallet/MyEtherWallet/pull/1722]
 - Update packages from Greenkeeper [#1885](https://github.com/MyEtherWallet/MyEtherWallet/pull/1885)
 
 ### Release v5.2.11